+++ components/tutors-ui/lib/Atoms/Iconography/themes.ts
@@ -1,5 +1,5 @@
-import { FluentIconLib } from "./support/fluent-icons";
-import type { IconType } from "../types/icon-types";
+import { FluentIconLib } from './support/fluent-icons';
+import type { IconType } from '../../../../tutors-reader-lib/src/types/icon-types';
 
 let StandardIconLib = FluentIconLib;
 
@@ -7,14 +7,20 @@ export function setIconLib(iconLib: any) {
   StandardIconLib = iconLib;
 }
 
-export const themes = ["tutors", "tutors-dark", "tutors-black", "tutors-dyslexia", "tutors-wireframe"];
+export const themes = [
+  'tutors',
+  'tutors-dark',
+  'tutors-black',
+  'tutors-dyslexia',
+  'tutors-wireframe'
+];
 
 export const themeIcons = {
   tutors: FluentIconLib,
-  "tutors-dark": FluentIconLib,
-  "tutors-dyslexia": FluentIconLib,
-  "tutors-black": FluentIconLib,
-  "tutors-wireframe": FluentIconLib,
+  'tutors-dark': FluentIconLib,
+  'tutors-dyslexia': FluentIconLib,
+  'tutors-black': FluentIconLib,
+  'tutors-wireframe': FluentIconLib
 };
 
 export function getIcon(type: string): IconType {
