+++ sites/tutors-time-reader/src/pages/support/CalendarTime.svelte
@@ -3,7 +3,7 @@
   import { Grid } from "ag-grid-community";
   import { CalendarSheet, options } from "../../components/sheets/calendar-sheet";
   import type { CourseService } from "../../reader-lib/services/course-service";
-  import Icon from "tutors-reader-lib/src/iconography/Icon.svelte";
+  import Icon from "tutors-ui/lib/Atoms/Iconography/Icon.svelte";
   import type { MetricsService } from "src/reader-lib/services/metrics-service";
 
   export let id;
